merged: changes from trunk. revs: r24197-r24308
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/branches/pvr-testing@24311 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ XBMC/xbmc/linux/DeviceKitDisksProvider.h
@@ -40,7 +40,7 @@ public:
 
   CStdString m_UDI, m_DeviceKitUDI, m_MountPath, m_FileSystem, m_Label;
   bool m_isMounted, m_isMountedByUs, m_isRemovable, m_isPartition;
-  long int m_PartitionSizeGiB;
+  float m_PartitionSizeGiB;
 };
 
 class CDeviceKitDiskDeviceNewAPI : public CDeviceKitDiskDevice
